meego-quality-automation/requirements.txt
PyYAML>=6.0
requests>=2.31.0
colorama>=0.4.6